feat: implement image upload and management for chefs
- Added file upload capability in ChefsController using Multer.
- Configured Multer storage to save uploaded images in a designated directory, creating it if it doesn't exist.
- Updated ChefsService to handle image paths for newly created and updated chefs, including deletion of old images.
- Enhanced create and update methods to accept and process image files.

+++ epicure-server/src/modules/chefs/chefs.service.ts
@@ -1,10 +1,12 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Body, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateChefDto } from './dto/create-chef.dto';
 import { InjectModel } from '@nestjs/mongoose';
 import mongoose, { Types } from 'mongoose';
 import { Chef } from './schemas/chef.schema';
 import { UpdateChefDto } from './dto/update-chef.dto';
 import { Restaurant } from '../restaurants/schemas/restaurant.schema';
+import * as fs from 'fs';
+import * as path from 'path';
 
 @Injectable()
 export class ChefsService {
@@ -15,7 +17,40 @@ export class ChefsService {
     private restaurantModel: mongoose.Model<Restaurant>,
   ) {}
 
-  async create(createChefDto: CreateChefDto): Promise<Chef> {
+  uploadImage(file: Express.Multer.File): string {
+    if (!file) {
+      throw new Error('No file uploaded');
+    }
+    return `uploads/chefs/${file.filename}`;
+  }
+
+  private deleteImageFile(imagePath: string) {
+    try {
+      const normalizedPath = imagePath.replace(/^\.\//, '');
+      const projectRoot = path.join(__dirname, '..', '..', '..', '..');
+      const absolutePath = path.join(projectRoot, 'public', normalizedPath);
+
+      console.log('Trying to delete file at:', absolutePath);
+
+      if (fs.existsSync(absolutePath)) {
+        fs.unlinkSync(absolutePath);
+        console.log('Successfully deleted file');
+      } else {
+        console.log('File does not exist');
+      }
+    } catch (error) {
+      console.error('Error deleting image file:', error);
+    }
+  }
+
+  async create(
+    @Body() createChefDto: CreateChefDto,
+    file?: Express.Multer.File,
+  ): Promise<Chef> {
+    if (file) {
+      const imagePath = this.uploadImage(file);
+      createChefDto.imgUrl = imagePath;
+    }
     const chef = await this.chefModel.create(createChefDto);
     return chef;
   }
@@ -36,12 +71,27 @@ export class ChefsService {
   async update(
     id: Types.ObjectId,
     updateChefDto: UpdateChefDto,
+    file?: Express.Multer.File,
   ): Promise<Chef> {
+    const currentChef = await this.chefModel.findById(id);
+    if (!currentChef) {
+      throw new NotFoundException('Chef not found');
+    }
+
+    if (file) {
+      if (currentChef.imgUrl) {
+        this.deleteImageFile(currentChef.imgUrl);
+      }
+      const imagePath = this.uploadImage(file);
+      updateChefDto.imgUrl = imagePath;
+    }
+
     const updatedChef = await this.chefModel.findByIdAndUpdate(
       id,
       updateChefDto,
       { new: true },
     );
+
     if (!updatedChef) {
       throw new NotFoundException('Chef not found');
     }
@@ -49,7 +99,15 @@ export class ChefsService {
   }
 
   async remove(id: Types.ObjectId): Promise<Chef> {
-    // First, check if the chef exists and delete them
+    const chef = await this.chefModel.findById(id);
+    if (!chef) {
+      throw new NotFoundException('Chef not found');
+    }
+
+    if (chef.imgUrl) {
+      this.deleteImageFile(chef.imgUrl);
+    }
+
     const deletedChef = await this.chefModel.findByIdAndDelete(id);
     if (!deletedChef) {
       throw new NotFoundException('Chef not found');
